refactor: reorganize form components into a dedicated Forms directory
