+++ src/sections/auth/components/Forms/recovery-form.tsx
@@ -5,8 +5,8 @@ import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { cn } from "@/lib/utils";
 import LucideIcon from "@/components/lucide-icon-name";
-import useRecovery from "../hooks/use-recovery";
-import { recoverySchema } from "../schemas/recovery";
+import useRecovery from "../../hooks/use-recovery";
+import { recoverySchema } from "../../schemas/recovery";
 
 export const RecoveryForm = () => {
   const { handleSubmit, isLoading } = useRecovery();
