+++ src/sections/auth/views/login-view.tsx
@@ -1,11 +1,14 @@
 import React from "react";
-import { LoginForm } from "../components/login-form";
+import { LoginForm } from "../components/Forms/login-form";
 import { AuthCard } from "../components/auth-card";
 
 export const LoginView: React.FC = () => {
   return (
     <>
-      <AuthCard title="Bienvenido" description="Por favor inicia sesión en tu cuenta">
+      <AuthCard
+        title="Bienvenido"
+        description="Por favor inicia sesión en tu cuenta"
+      >
         <LoginForm />
       </AuthCard>
     </>
